refactor: Split up components
